+++ example/data_source_disk.tf
@@ -1,6 +1,6 @@
 data "clouddk_disk" "example" {
-  id = "${element(data.clouddk_disks.example.ids, 0)}"
-  server_id = "${element(data.clouddk_servers.example.ids, 0)}"
+  id        = "${element(flatten(clouddk_server.example.disk_ids), 0)}"
+  server_id = "${clouddk_server.example.id}"
 }
 
 output "data_clouddk_disk_example_label" {