+++ resource_server.go
@@ -386,22 +386,48 @@ func resourceServerUpdate(d *schema.ResourceData, m interface{}) error {
 func resourceServerDelete(d *schema.ResourceData, m interface{}) error {
 	clientSettings := m.(ClientSettings)
 
-	req, reqErr := getClientRequestObject(&clientSettings, "DELETE", fmt.Sprintf("cloudservers/%s", d.Id()), new(bytes.Buffer))
+	// Due to some locking issues caused by how cloud.dk works, we may initially receive an HTTP 500 response. In this case we want to
+	// try to delete the resource again once we have waited for a short period of time. We do this multiple times before giving up.
+	timeDelay := int64(10)
+	timeMax := float64(600)
+	timeStart := time.Now()
+	timeElapsed := timeStart.Sub(timeStart)
 
-	if reqErr != nil {
-		return reqErr
+	responseStatus := ""
+	serverHostname := d.Get(ResourceServerHostname).(string)
+
+	for timeElapsed.Seconds() < timeMax {
+		if int64(timeElapsed.Seconds())%timeDelay == 0 {
+			log.Printf("[DEBUG] Querying the API in order to delete server '%s' (id: %s)", serverHostname, d.Id())
+
+			req, reqErr := getClientRequestObject(&clientSettings, "DELETE", fmt.Sprintf("cloudservers/%s", d.Id()), new(bytes.Buffer))
+
+			if reqErr != nil {
+				return reqErr
+			}
+
+			client := &http.Client{}
+			res, resErr := client.Do(req)
+
+			if resErr != nil {
+				return resErr
+			} else if res.StatusCode == 200 || res.StatusCode == 404 {
+				d.SetId("")
+
+				return nil
+			} else {
+				responseStatus = res.Status
+
+				log.Printf("[DEBUG] Failed to delete server '%s' (id: %s) - Retrying in %d seconds", serverHostname, d.Id(), timeDelay)
+			}
+
+			time.Sleep(1 * time.Second)
+		}
+
+		time.Sleep(100 * time.Millisecond)
+
+		timeElapsed = time.Now().Sub(timeStart)
 	}
 
-	client := &http.Client{}
-	res, resErr := client.Do(req)
-
-	if resErr != nil {
-		return resErr
-	} else if res.StatusCode != 200 && res.StatusCode != 404 {
-		return fmt.Errorf("Failed to delete the server - Reason: The API responded with HTTP %s", res.Status)
-	}
-
-	d.SetId("")
-
-	return nil
+	return fmt.Errorf("Failed to delete the server - Reason: The API responded with HTTP %s", responseStatus)
 }