+++ CHANGELOG.md
@@ -14,4 +14,5 @@ FEATURES:
 * **New Data Source:** `clouddk_server`
 * **New Data Source:** `clouddk_servers`
 * **New Data Source:** `clouddk_templates`
+* **New Resource:** `clouddk_firewall_rule`
 * **New Resource:** `clouddk_server`